chore: lib refactoring
