+++ Cargo.toml
@@ -17,7 +17,6 @@ exclude = [
     ".github/**",
     "tests/**",
     ".gitignore",
-    "todo.txt"
 ]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -30,17 +29,17 @@ required-features = ["cli"]
 chumsky = { version = "0.8.0", default-features = false }
 
 [dependencies.textwrap]
-version = "0.15.0"
+version = "0.15.1"
 default-features = false
 # optional = true
 
 [dependencies.comfy-table]
-version = "6.0.0"
+version = "6.1.0"
 default-features = false
 # optional = true
 
 [dependencies.lexopt]
-version = "0.2.0"
+version = "0.2.1"
 optional = true
 
 [features]
