+++ src/parser/tags/mod.rs
@@ -14,3 +14,13 @@ mod r#type;
 pub use r#type::*;
 mod module;
 pub use module::*;
+mod see;
+pub use see::*;
+mod usage;
+pub use usage::*;
+
+#[derive(Debug, Default, Clone)]
+pub struct Prefix {
+    pub left: Option<String>,
+    pub right: Option<String>,
+}
