--- todo.txt
@@ -1,30 +1,0 @@
-## TODO
-
-[x] optional description
-[x] simple comments (bcz comments can appear before @class)
-[x] optional parameters (but IDK how to render is nicely)
-[ ] expand `Ty`
-[x] description in return
-[x] consider 80 char width
-[x] @usage/@example
-[x] multiline @param description
-[x] multiline @field description
-[x] multiple @see reference
-[x] func w/ no returns or no params
-[x] return with only type
-[x] support metamethods i.e. this:that()
-[x] check for exported `return`
-[x] enum via @alias
-[ ] also support string as an entry alongside path
-[x] long names
-[x] Non-Standard tags (should not interfere with LSP)
-    [x] ---@module <tag> <description>
-            (override the tagged name prefix like *U.comment* -> *comment-comment*)
-    [x] ---@divider <char>
-    [x] ---@private
-    [x] ---@toc <tag>
-[ ] Error handling
-
-## Dropped
-
-[ ] @vararg
